jack: removed sample_rate callback
Currently, the JACK plugin manipulates the audio_format struct which
was passed to the open() method.  This is very likely to break,
because the plugin must not permanently store this pointer.  After
this patch, MPD ignores sample rate changes.  It looks like other
software is doing the same, and I guess this is a non-issue.

This patch converts the audio_format pointer within jack_data into a
static audio_format struct.