configure.ac: use /Developer/SDKs/MacOSX10.5.sdk
First, this is not a nice fix. I am also not sure why the error
happens in the first place. I assume Apple deprecated some stuff in
10.6 + x86_64.

My patch simply uses the 10.5 SDK if compiling on 10.6 Snow Leopard -
which is installed by default if you install XCode 3.2 that comes with
Snow Leopard. The reason this is not a nice fix is, of course, that
this doesn't fix the problem. It just "postpones" it to the next
release of MacOS X. But, some people may need it, and its better than
nothing.