PlayerThread: Only drop 0 length packets without tags
Fixes a regression from 752dfb3d95482c562e5d24c6ea839c4815de9a6d which
caused the current chunk to be flushed as soon as new replaygain
information was found. If this occurs on a tag chunk, it has no data
(length 0) and is then skipped before pushing it to all of the outputs.

This change allows 0-length chunks through if they contain a tag and
they are now appearing in mplayer and mpv properly.