configure.ac: check for pthread before librt
Works around a linker problem with some older glibc versions: if
"-lrt" was used, then "-pthread" was implied, but only the symbols
used by librt were available.  This led to a linker error because
pthread_atfork() was not found.

So with "-lrt", autoconf decides that no pthread flag is necessary,
but in the end fails due to pthread_atfork() missing.

By checking for pthread before librt, we avoid this dependency
problem.